rename forest.view to forest.map_view
