+++ forest/drivers/ghrsstl4.py
@@ -19,7 +19,7 @@ except ModuleNotFoundError:
 
 import glob
 from forest import geo
-from forest.view import UMView
+from forest.map_view import UMView
 from forest.util import to_datetime as _to_datetime
